01.02.2023 Refaktor crawlera - hermetyzacja logiki odpowiadającej za …

+++ GPU Scraper/Entities/GPU.cs
@@ -4,9 +4,9 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double LowestPrice { get; set; }
+        public decimal LowestPrice { get; set; }
         public string LowestPriceShop { get; set; }
-        public double? HighestPrice { get; set; }
+        public decimal? HighestPrice { get; set; }
         public string? HighestPriceShop { get; set; }
     }
 }
