+++ GPUScraper.Models/Models/GPUDto.cs
@@ -9,9 +9,9 @@ namespace GPUScraper.Models.Models
     public class GPUDto
     {
         public string Name { get; set; }
-        public double LowestPrice { get; set; }
+        public decimal LowestPrice { get; set; }
         public string LowestPriceShop { get; set; }
-        public double? HighestPrice { get; set; } = null;
+        public decimal? HighestPrice { get; set; } = null;
         public string? HighestPriceShop { get; set; } = null;
     }
 }
